pyproject.toml
[project]
name = "hurr-ai"
version = "0.1.0"
description = "Multimodal embedding benchmarking web application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.1.0",
    "peewee>=3.17.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "httpx>=0.28.0",
    "cohere>=5.0.0",
    "Pillow>=11.0.0",
    "numpy>=2.0.0",
]

[dependency-groups]
dev = []
